Worked on CreateGame
I worked on the styling of the page, which is mostly done. I made sure the page is updated correctly upon receiving user input. I also tried to ensure that the proper message is sent to the back-end (I couldn't verify whether that worked). And most importantly I inserted a beautiful (I repeat: beautiful) JustOne logo...